Added sample codes and QR codes
These are the codes used during TechVenture 2k24, and corresponding QR codes.
You may use the same QR codes to test the working of the app, and also for next event, just make sure that the order in which tehy appear doesn't change.
OR you may generate new QR codes from the website given in `Readme.md` file.
But develop new code snippets for the new event, to make it more interesting.
Make the necessary changes in the code to make it work with the new event.
You can also add new features as discussed already.